refactor: use extensible attribution naming
