--- packages/kit/src/components/LastActivityTracker/googlePlayInstallAttribution.ts
@@ -1,3 +1,0 @@
-export function reportGooglePlayInstallAttribution(): Promise<void> {
-  return Promise.resolve();
-}
