+++ packages/kit/src/components/LastActivityTracker/installAttribution.ts
@@ -1,0 +1,3 @@
+export function reportInstallAttribution(): Promise<void> {
+  return Promise.resolve();
+}
